requirements.txt
bottle==0.12.19
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
gitdb==4.0.7
GitPython==3.1.14
idna==2.10
jwt==1.2.0
pycparser==2.20
PyJWT==2.0.1
requests==2.25.1
smmap==4.0.0
urllib3==1.26.4
